// src/common/enums.ts
export enum eEmployeeState {
    ADDED = "ADDED",
    IN_CHECK = "IN_CHECK",
    APPROVED = "APPROVED",
    ACTIVE = "ACTIVE",
    INACTIVE = "INACTIVE",
}

export enum eStatusColors {
    YELLOW = "yellow",
    GREEN = "green",
    RED = "red",
}

export enum ASYNC_STATUS {
    IDLE = "idle",
    LOADING = "loading",
    FAILED = "failed",
}

export enum eTransitionDirection {
    NEXT = "NEXT",
    PREVIOUS = "PREVIOUS",
}